authentication works for delete, create, and edit

+++ server/controllers/posts.js
@@ -11,9 +11,12 @@ export const getPosts = async (req, res) => {
 }
 
 export const createPost = async (req, res) => {
-    const body = req.body;
+    const body = req.body
+
+    if(!req.userId) return res.json({message: "Unauthenticated"})
+    const UID = req.userId
     try {
-        const newPost = await PostMessage.create(body)
+        const newPost = await PostMessage.create({...body, UID: UID})
         res.status(200).json(newPost)
     } catch (error) {
         res.status(409).json({message: err.message})
@@ -22,8 +25,11 @@ export const createPost = async (req, res) => {
 
 export const deletePost = async (req, res) => {
     const {id} = req.body
+
+    if(!req.userId) return res.json({message: "Unauthenticated"})
+    const UID = req.userId
     try {
-        const post = await PostMessage.findOneAndDelete({_id: id})
+        const post = await PostMessage.findOneAndDelete({UID: UID, _id: id})
         res.status(200).json(post)
     } catch(err) {
         res.status(404).json({message: err.message})
@@ -32,12 +38,17 @@ export const deletePost = async (req, res) => {
 
 export const updatePost = async (req, res) => {
     const {id} = req.params
+
+    if(!req.userId) return res.json({message: "Unauthenticated"})
+    const UID = req.userId
+
     const body = req.body
+    console.log(body)
     if (!mongoose.Types.ObjectId.isValid(id)){
         return res.status(404).send("No post with that id")
     }
     try {
-        const post = await PostMessage.findOneAndUpdate({_id: id}, body, {new: true})
+        const post = await PostMessage.findOneAndUpdate({_id: id, UID: UID}, body, {new: true})
         res.status(200).json(post)
     } catch(err) {
         res.status(404).json({message: err.message})
@@ -45,9 +56,27 @@ export const updatePost = async (req, res) => {
 }
 
 export const likePost = async (req, res) => {
+    const {id} = req.params;
+
+    if(!req.userId) return res.json({message: "Unauthenticated"})
+
     try {
+        if (!mongoose.Types.ObjectId.isValid(id)) return res.status(404).send("No post with this id")
         
+        const post = await PostMessage.findById(id)
+
+        const index = post.likeCount.findIndex((id) => id === String(req.userId))
+
+        if (index === -1){
+            post.likeCount.push(req.userId)
+        } else {
+            post.likeCount.filter((id) => id !== String(req.userId))
+        }
+
+        const updatedPost = await PostMessage.findByIdAndUpdate(id, post, {new: true})
+
+        res.json(updatedPost)
     } catch (error) {
-        
+        res.status(404).json({message: err.message})
     }
 }
